docs: refactor ExamplePreview components
