+++ docs/.vuepress/example-preview/utils.js
@@ -1,0 +1,26 @@
+import { getParameters } from 'codesandbox/lib/api/define'
+
+export function getCodesandboxParameters(files) {
+  const data = {
+    files: {
+      'package.json': {
+        content: {
+          main: 'main.js',
+          dependencies: {
+            vue: 'latest',
+            'vue-router': 'latest'
+          }
+        }
+      },
+      ...files.reduce(
+        (fileMap, file) => ({
+          ...fileMap,
+          ...file
+        }),
+        {}
+      )
+    }
+  }
+
+  return getParameters(data)
+}
